Move terminal state impl to separate module
